requirements.txt
websocket_client
mecab-python
markovify
python-dotenv
mecab
